Organize regression tests by module
